+++ src/main/java/cn/cash/register/dao/MemberInfoMapper.java
@@ -7,7 +7,7 @@ public interface MemberInfoMapper {
 
     int insert(MemberInfo record);
 
-    int insertSelective(MemberInfo record);
+    Long insertSelective(MemberInfo record);
 
     MemberInfo selectByPrimaryKey(Long id);
